backend/reporting_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
minio
python-jose[cryptography]
httpx
rq
redis
nbformat
jupyter_client
